Handle importing the various types of samples files
We now allow samples files tobe structured as follows:

1. <directory>/docs, <directory>/design_docs

This is the recommended setup. You should have a single folder that
contains a docs and a design_docs folder in it.

2. /docs, /design_docs

This should be considered a deprecated format. If means that the zip
file contains two directories. One is called docs and the other is
called design_docs.

3. /

This should definitely not be used for new samples zip files, but it
is supported because we have some old samples zip folders that use it.
This format just has a bunch of json document files in the root directory
of the zip file.

Change-Id: Ia42d2524fc4a03a958c47da4d393d6e3663606ee
Reviewed-on: http://review.couchbase.org/77756
Reviewed-by: Michael Wiederhold <mike@couchbase.com>
Tested-by: Michael Wiederhold <mike@couchbase.com>