MB-59678 Remove max TTL field from manifest
CE does not allow collections to have a max TTL which was causing
cbimport to fail on the travel-sample, as we were trying to submit them.
To fix we omit the field, which will stop the field from being sent.
This means that the sample will have the bucket TTL for all collections.

Change-Id: If01e2120971d142cf42fbb0b7a610b8114bd4a9f
Reviewed-on: https://review.couchbase.org/c/docloader/+/201032
Reviewed-by: James Lee <james.lee@couchbase.com>
Tested-by: Matt Hall <matt.hall@couchbase.com>
Reviewed-by: Nour Elmenyawi <nour.elmenyawi@couchbase.com>